requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
docker==7.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.1.0
gunicorn==21.2.0 